setup gatling prefab, add to charger, add layers, update some scripts
Added Player, Player Hitbox, Player Hurtbox, Enemy Hitbox, Enemy Hurtbox layers, to be used with projectiles. Updated weapon script to fire projectiles from the assigned fire point, and setup the proper layer/damage for the projectile.
Updated Projectile script to be able to work with a velocity, so it makes it less likely to run into your projectiles