Compile current assets for first playable
Charger and semi can drive through the areas of the level, the charger is controlled by the player, the semi is controlled by AI. The semi might still flip/freak out, but it is able to drive the entire length. You can fire the weapons on the charger by pressing left click.

Added mixer for audio, and updated prefab audio sources with the mixer
Made charger engine sound quieter, cause it was LOUD
Added RCC layer, car controller needed it
Added Playable 1 scene, lighting info, and navmesh
Added terrain meshes to the project